requirements.txt
requests
supabase
python-dotenv
google-generativeai
matplotlib